Finalizado details.

+++ api/app/Http/Controllers/Admin/PlanController.php
@@ -58,6 +58,12 @@ class PlanController extends Controller
         if (!$plan) {
             return redirect()->back();
         }
+        
+        if ($plan->details()->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'Existem detalhes vinculados a esse plano, portanto não pode deletar.');
+        }
 
         $plan->delete();
